refactor: Break server start out into own file / function for integration tests
